New RDPF2of3 struct
When computational peers do an ORAM read, they hold an RDPFTriple, but
they only need two of them.  Previously, we were moving the two RDPFs
they needed out of the RDPFTriple into an RDPFPair, breaking the triple.
That _was_ fine, but now that we're going to be reusing RDPFTriples when
performing related-index ORAM accesses, we can't break that triple.

RDPF2of3 acts just like an RDPFPair, except it holds const references to
two RDPFs (contained in an RDPFTriple), rather than holding the RDPFs
themselves.