New Shape::init() functions that initialize the contents of a Shape
You can pass either a constant (all elements of the shape are
initialized to that constant) or a function f : size_t -> size_t
(element i is initialized to f(i) for each i).  The function f must be
deterministic and public.

The initialization is entirely a local operation; no communication is
done.